* fhandler_socket.cc (fhandler_socket::send_internal): Change type of 	local variable res to ssize_t. 	* net.cc (cygwin_sendto): Fix local variable res, containing return 	value, to be ssize_t as well. 	(cygwin_recvfrom): Ditto. 	(cygwin_recv): Ditto. 	(cygwin_send): Ditto. 	(cygwin_recvmsg): Ditto. 	(cygwin_sendmsg): Ditto. 	* syscalls.cc (write): Ditto. 	(writev): Ditto. 	* pipe.cc (fhandler_pipe::open): Fix type when reading handle value to 	accommodate 32 bit compiler.

+++ winsup/cygwin/fhandler_socket.cc
@@ -1617,7 +1617,7 @@ fhandler_socket::recvmsg (struct msghdr *msg, int flags)
 inline ssize_t
 fhandler_socket::send_internal (struct _WSAMSG *wsamsg, int flags)
 {
-  int res = 0;
+  ssize_t res = 0;
   DWORD ret = 0, err = 0, sum = 0, off = 0;
   WSABUF buf;
   bool use_sendmsg = false;
